Working DAWGDic wrappers

+++ src/qac_impl/dawgdic/include/dawgdic/base-types.h
@@ -10,10 +10,10 @@ typedef char CharType;
 typedef unsigned char UCharType;
 
 // 32-bit integer.
-typedef long  ValueType;
+typedef int ValueType;
 
 // 32-bit unsigned integer.
-typedef unsigned long  BaseType;
+typedef unsigned int BaseType;
 
 // 32 or 64-bit unsigned integer.
 typedef std::size_t SizeType;
